django_bs/requirements.txt
Django==1.8
Landscape-Client==14.12
MySQL-python==1.2.3
PAM==0.4.2
South==1.0.2
Twisted-Core==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
colorama==0.2.5
configobj==4.7.2
djangorestframework==3.1.1
html5lib==0.999
pyOpenSSL==0.13
pyserial==2.6
python-apt==0.9.3.5ubuntu1
python-debian==0.1.21-nmu2ubuntu2
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
urllib3==1.7.1
wsgiref==0.1.2
zope.interface==4.0.5
